refactor(LottoWinningStatistics): 총 상금 계산 타입을 int → long으로 변경
- 총 상금(totalPrize)을 int에서 long으로 변경하여 큰 금액 계산 시 오버플로우 방지

+++ src/main/java/lotto/model/LottoWinningStatistics.java
@@ -22,14 +22,14 @@ public class LottoWinningStatistics {
     }
 
     public double calculateProfitRate() {
-        int totalPrize = calculateTotalPrize();
+        long totalPrize = calculateTotalPrize();
         return (double) totalPrize / lottoPurchaser.getPurchaseAmount() * 100;
     }
 
-    private int calculateTotalPrize() {
-        int totalPrize = 0;
+    private long calculateTotalPrize() {
+        long totalPrize = 0;
         for (Rank rank : winningCount.keySet()) {
-            totalPrize += rank.getPrize() * winningCount.put(rank, winningCount.getOrDefault(rank, 0));
+            totalPrize += (long) rank.getPrize() * winningCount.put(rank, winningCount.getOrDefault(rank, 0));
         }
         return totalPrize;
     }
